requirements.txt
requests>=2.31.0
lxml>=4.9.0
google-genai>=1.18.0
tweepy>=4.14.0
python-dotenv>=1.0.0
